update mip structure, remove commented out code in unet, and fix flake8 errors
